Use smaller units for hashrates and timestamps

+++ main/tasks/hashrate_monitor_task.h
@@ -5,7 +5,7 @@
 
 typedef struct {
     uint32_t value;
-    uint64_t time_ms;
+    uint32_t time_ms;
     float hashrate;
 } measurement_t;
